functions/requirements.txt
azure-functions
requests
pillow
azure-cognitiveservices-vision-customvision
azure-storage-blob